Delete .vscode directory
This may actually be the source of the "math.h" not found issue.
My theory is that, because this is a hidden file, its meant to be stored locally on everyones computers with their own versions and configs instead of one copy on git. Since dropbox doesn't know that though, it assigns one .vscode to everybody, causing conflict issues whenever it is changed. 
We'll have to do more experiments on how this functions with git but this might be the solution to that problem